Reverted Thread receiving loop adaptions back to original code as the current code caused hangs and incomplete messages when multiple messages are received. Instead ensured that all XML-Packets are read with surelyRead function like in original code which handles multiple messages in receive-buffer correctly. Added dynamic memory management to surelyRead.
git-svn-id: http://svn.miranda-ng.org/main/branches/msnp24@13568 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ src/msn_global.h
@@ -550,7 +550,7 @@ struct ThreadData
    //----| internal data buffer |--------------------------------------------------------
    int           mBytesInData;     // bytes available in data buffer
    char          *mData;           // data buffer for connection
-   int           mDataSize;
+   size_t        mDataSize;
 
    //----| methods |---------------------------------------------------------------------
    void          applyGatewayData(HANDLE hConn, bool isPoll);
